% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_swuds}
\alias{aggregate_swuds}
\alias{aggregate_season}
\title{aggregate functions}
\usage{
aggregate_swuds(s_wuds, filter_by, filter_val = NA,
  aggregate_by = "Volume_mgd")

aggregate_season(s_wuds, season, sum_avg)
}
\arguments{
\item{s_wuds}{swuds object (data frame from melt function)}

\item{season}{chr, the aquifer of interest to be aggregated}

\item{sum_avg}{chr, user selects either "sum" or "avg" and the sum
or average will be computed}
}
\description{
Allows user to aggregate their data from SWUDS by user-defined parameters.
}
\examples{
s_wuds <- swuds_sample #example data from Ohio
test_all_huc <- aggregate_swuds(s_wuds, "FROM_HUC_CD")
test_some_huc <- aggregate_swuds(s_wuds, "FROM_HUC_CD", c("041000040104", "050600020105"))
test_some_years <- aggregate_swuds(s_wuds, "YEAR", c("1995", "2004"))
#test_some_aquifers <- aggregate_swuds(s_wuds, "aquifer", "TBD")
test_some_months <- aggregate_swuds(s_wuds, "Month", c("Jan", "Feb"))
test_some_state_county <- aggregate_swuds(s_wuds, "state_county", "39033")
test_some_seasons <- aggregate_swuds(s_wuds, "season", "Fall")
season <- "Fall"
test_aggSeason <- aggregate_season(s_wuds, season, sum_avg)
}
