% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_aquifer.R
\name{aggregate_aquifier}
\alias{aggregate_aquifier}
\title{aggregate_HUC}
\usage{
aggregate_aquifier(s.wuds, aquifer, sum_avg)
}
\arguments{
\item{s.wuds}{dataframe, the swuds water use data}

\item{aquifer}{chr, the aquifer of interest to be aggregated}

\item{sum_avg}{chr, user selects either "sum" or "avg" and the sum or average will be computed}
}
\description{
Allows user to aggregate their data from SWUDS by user-defined aquifer(s)
}
\examples{
s.wuds <- swuds_sample #example data from Ohio
aquifer <- "TBD"
sum_avg <- "sum"
test_aquifer <- aggregate_aquifier(s.wuds, aquifer, sum_avg)

}
