% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_HUC.R, R/aggregate_aquifer.R
\name{aggregate_HUC}
\alias{aggregate_HUC}
\title{aggregate_HUC}
\usage{
aggregate_HUC(s.wuds, aquifer, sum_avg)

aggregate_HUC(s.wuds, aquifer, sum_avg)
}
\arguments{
\item{s.wuds}{dataframe, the swuds water use data}

\item{aquifer}{chr, the aquifer of interest to be aggregated}

\item{sum_avg}{chr, user selects either "sum" or "avg" and the sum or average will be computed}

\item{HUC}{chr, the HUC of interest to be aggregated}

\item{s.wuds}{dataframe, the swuds water use data}

\item{sum_avg}{chr, user selects either "sum" or "avg" and the sum or average will be computed}
}
\description{
Allows user to aggregate their data from SWUDS by user-defined HUC(S)

Allows user to aggregate their data from SWUDS by user-defined aquifer(s)
}
\examples{
s.wuds <- swudsSample #example data from Ohio
HUC <- c("041000040104", "050600020105")
sum_avg <- "sum"
test_HUC <- aggregate_HUC(s.wuds, HUC, sum_avg)

s.wuds <- swudsSample #example data from Ohio
aquifer <- TBD
sum_avg <- "sum"
test_aquifer <- aggregate_aquifer(s.wuds, aquifer, sum_avg)

}
