% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_month.R
\name{aggregate_month}
\alias{aggregate_month}
\title{aggregate_month}
\usage{
aggregate_month(s_wuds, month, sum_avg)
}
\arguments{
\item{s_wuds}{dataframe, the swuds water use data}

\item{month}{chr, the month of interest to be aggregated in abb format (i.e. Jan, Feb, etc.)}

\item{sum_avg}{chr, user selects either "sum" or "avg" and the sum or average will be computed}
}
\description{
Allows user to aggregate their data from SWUDS by user-defined month(s)
}
\examples{
s_wuds <- swuds_sample #example data from Ohio
month <- c("Jan", "Feb")
sum_avg <- "sum"
test_month <- aggregate_month(s_wuds, month, sum_avg)
}
