% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_season.R
\name{aggregate_season}
\alias{aggregate_season}
\title{aggregate_season}
\usage{
aggregate_season(s.wuds, season, sum_avg)
}
\arguments{
\item{s.wuds}{dataframe, the swuds water use data}

\item{season}{chr, the aquifer of interest to be aggregated}

\item{sum_avg}{chr, user selects either "sum" or "avg" and the sum or average will be computed}
}
\description{
Allows user to filter their water use data from SWUDS by season(s)
}
\examples{
s.wuds <- swuds_sample #example data from Ohio
sum_avg <- "sum"
season <- "Fall"
test_aggSeason <- aggregate_season(s.wuds, season, sum_avg)

}
