% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_year.R
\name{aggregate_year}
\alias{aggregate_year}
\title{aggregate_year}
\usage{
aggregate_year(s_wuds, year, sum_avg)
}
\arguments{
\item{s_wuds}{dataframe, the swuds water use data}

\item{year}{numeric, the year of interest to be aggregated}

\item{sum_avg}{chr, user selects either "sum" or "avg" and the sum or average will be computed}
}
\description{
Allows user to aggregate their data from SWUDS by user-defined year(s)
}
\examples{
s_wuds <- swuds_sample #example data from Ohio
year <- c("1995", "2004")
sum_avg <- "sum"
test_year <- aggregate_year(s_wuds, year, sum_avg)

}
