% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_swuds.R
\name{explore_swuds}
\alias{explore_swuds}
\title{Explore data in the Shiny Application}
\usage{
explore_swuds(browse = TRUE)
}
\arguments{
\item{browse}{Logical. Use browser for running Shiny app.}
}
\description{
Open an interactive app in a browser. See the "Shiny App" vignette:
\href{../doc/shinyApp.html}{\code{vignette("shinyApp", package = "toxEval")}} for more details. Using this 
function is a quick and convenient way
to explore data. For more customization, the R-code to 
produce each graph and table is displayed in the app. That is 
a good starting-point for a custom analysis.
}
\examples{
\dontrun{
explore_swuds()
}
}
