% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_aquifer.R
\name{filter_aquifer}
\alias{filter_aquifer}
\title{filter_aquifer}
\usage{
filter_aquifer(s_wuds, aquifer)
}
\arguments{
\item{s_wuds}{dataframe, the swuds water use data}

\item{aquifer}{chr, the aquifer of interest to be filtered}
}
\description{
Allows user to filter their water use data from SWUDS by aquifer
}
\examples{
s_wuds <- swuds_sample #example data from Ohio
aquifer <- "TBD"
test_aquifer <- filter_aquifer(s_wuds, aquifer)

}
