% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_category.R
\name{filter_category}
\alias{filter_category}
\title{filter_category}
\usage{
filter_category(s.wuds, category)
}
\arguments{
\item{s.wuds}{dataframe, the swuds water use data}

\item{category}{chr, the category of interest to be filtered}
}
\description{
Allows user to filter their water use data from SWUDS by category
}
\examples{
s.wuds <- swuds_sample #example data from Ohio
category <- c("WS", "IR", "IN")
test_category <- filter_category(s.wuds, category)

}
