% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_county.R
\name{filter_county}
\alias{filter_county}
\title{filter_county}
\usage{
filter_county(s_wuds, state_county)
}
\arguments{
\item{s_wuds}{dataframe, the swuds water use data}

\item{state_county}{chr, the county of interested to be filtered.
The state abb is appended to the front of the county name}
}
\description{
Allows user to filter their water use data from SWUDS by county
}
\examples{
s_wuds <- swuds_sample #example data from Ohio
state_county <- c("Ohio_Madison County", "Ohio_Licking County")
test_county <- filter_county(s_wuds, state_county)
}
