% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_HUC.R
\name{filter_HUC}
\alias{filter_HUC}
\title{filter_HUC}
\usage{
filter_HUC(s.wuds, HUC)
}
\arguments{
\item{s.wuds}{dataframe, the swuds water use data}

\item{HUC}{chr, the HUC of interest to be aggregated}
}
\description{
Allows user to filter their water use data from SWUDS by HUC
}
\examples{
s.wuds <- swudsSample #example data from Ohio
HUC <- c("041000040104", "050600020105")
test_HUC <- filter_HUC(s.wuds, HUC)

}
