% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_month.R
\name{filter_month}
\alias{filter_month}
\title{filter_month}
\usage{
filter_month(s_wuds, month)
}
\arguments{
\item{s_wuds}{dataframe, the swuds water use data}

\item{month}{chr, the month of interest to be filtered in abb format (e.g. Jan, Feb, etc.)}
}
\description{
Allows user to filter their water use data from SWUDS by month(s)
}
\examples{
s_wuds <- swuds_sample #example data from Ohio
month <- c("Jan", "Feb")
test_month <- filter_month(s_wuds, month)

}
