% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_siteType.R
\name{filter_site_type}
\alias{filter_site_type}
\title{filter_site_type}
\usage{
filter_site_type(s_wuds, site_type)
}
\arguments{
\item{s_wuds}{dataframe, the swuds water use data}

\item{site_type}{chr, the site type of intesitest to be filtered}
}
\description{
Allows user to filter their water use data from SWUDS by site type(s)
}
\examples{
s_wuds <- swuds_sample #example data from Ohio
site_type <- c("AS", "FA-DV")
test_site_type <- filter_site_type(s_wuds, site_type)
}
