% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_siteType.R
\name{filter_siteType}
\alias{filter_siteType}
\title{filter_siteType}
\usage{
filter_siteType(s.wuds, siteType)
}
\arguments{
\item{s.wuds}{dataframe, the swuds water use data}

\item{siteType}{chr, the site type of intesitest to be filtered}
}
\description{
Allows user to filter their water use data from SWUDS by site type(s)
}
\examples{
s.wuds <- swudsSample #example data from Ohio
siteType <- c("AS", "FA-DV")
test_siteType <- filter_siteType(s.wuds, siteType)

}
