% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_state.R
\name{filter_state}
\alias{filter_state}
\title{filter_state}
\usage{
filter_state(s_wuds, state)
}
\arguments{
\item{s_wuds}{dataframe, the swuds water use data}

\item{state}{chr, the state of interest to be filtered}
}
\description{
Allows user to filter their water use data from SWUDS by state(s)
}
\examples{
s_wuds <- swuds_sample #example data from Ohio
state <- c("Ohio")
test_state <- filter_state(s_wuds, state)
}
