% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_waterType.R
\name{filter_water_type}
\alias{filter_water_type}
\title{filter_water_type}
\usage{
filter_water_type(s_wuds, water_type)
}
\arguments{
\item{s_wuds}{dataframe, the swuds water use data}

\item{water_type}{chr, the water type of interest to be filtered}
}
\description{
Allows user to filter their water use data from SWUDS by water type(s)
}
\examples{
s_wuds <- swuds_sample #example data from Ohio
water_type <- "SW"
test_water_type <- filter_water_type(s_wuds, water_type)
}
