% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_waterType.R
\name{filter_waterType}
\alias{filter_waterType}
\title{filter_waterType}
\usage{
filter_waterType(s.wuds, waterType)
}
\arguments{
\item{s.wuds}{dataframe, the swuds water use data}

\item{waterType}{chr, the water type of interest to be filtered}
}
\description{
Allows user to filter their water use data from SWUDS by water type(s)
}
\examples{
s.wuds <- swudsSample #example data from Ohio
waterType <- "SW"
test_waterType <- filter_waterType(s.wuds, waterType)
}
