% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_year.R
\name{filter_year}
\alias{filter_year}
\title{filter_year}
\usage{
filter_year(s_wuds, year)
}
\arguments{
\item{s_wuds}{dataframe, the swuds water use data}

\item{year}{numeric, the year of interest to be filtered}
}
\description{
Allows user to filter their water use data from SWUDS by year(s)
}
\examples{
s_wuds <- swuds_sample #example data from Ohio
year <- c("1995", "2009", "2012")
test_year <- filter_year(s_wuds, year)
}
