% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swuds_parse_functions.R
\name{melt_water_quant_pop}
\alias{melt_water_quant_pop}
\title{melt_water_quant_pop}
\usage{
melt_water_quant_pop(df_in)
}
\arguments{
\item{df_in}{data frame}
}
\description{
Function to create month, year, month#, day, date, decimal date, and water
year to the mergeWaterQuantPopServe data frame
}
\examples{
pathToSample <- system.file("extdata",package = "WUReview")
dp <- read_swuds_pop(file.path(pathToSample,"OHpopserved_output.xlsx"))
dq <- read_swuds_quant(file.path(pathToSample,
              "OH_CTF_SW_monthly_permit_sample_data.xlsx"))
df <- merge_dq_dp(dq, dp)
df_melt <- melt_water_quant_pop(df)
}
