% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swuds_parse_functions.R
\name{meltWaterQuantPopServ}
\alias{meltWaterQuantPopServ}
\title{meltWaterQuantPopServ}
\usage{
meltWaterQuantPopServ(mergeWaterQuantPopServ)
}
\arguments{
\item{mergeWaterQuantPopServ}{data frame}
}
\description{
Function to create month, year, month#, day, date, decimal date, and water
year to the mergeWaterQuantPopServe data frame
}
\examples{
pathToSample <- system.file("extdata",package = "wateRuseSWUDS")
dp <- readPopServedXL(file.path(pathToSample,"OHpopserved_output.xlsx"))
dq <- readWaterQuantityXL(file.path(pathToSample,
              "OH_CTF_SW_monthly_permit_sample_data.xlsx"))
df <- mergeWaterQuantPopServ(waterQuantDF = dq, popServDF = dp)
df_melt <- meltWaterQuantPopServ(mergeWaterQuantPopServ = df)
}
