% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swuds_parse_functions.R
\name{merge_dq_dp}
\alias{merge_dq_dp}
\title{Merge Water Quantity and Population}
\usage{
merge_dq_dp(dq, dp)
}
\arguments{
\item{dq}{data frame}

\item{dp}{data frame}
}
\description{
Function to merge Water Quantity and Population Served data frames into one
by "to_agency code", "to_site_no", "Year"
}
\examples{
pathToSample <- system.file("extdata",package = "wateRuseSWUDS")
dp <- read_swuds_pop(file.path(pathToSample,"OHpopserved_output.xlsx"))
dq <- read_swuds_quant(file.path(pathToSample,
              "OH_CTF_SW_monthly_permit_sample_data.xlsx"))
df <- merge_dq_dp(dq = dq, dp = dp)
}
