% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swuds_parse_functions.R
\name{mergeWaterQuantPopServ}
\alias{mergeWaterQuantPopServ}
\title{Merge Water Quantity and Population}
\usage{
mergeWaterQuantPopServ(waterQuantDF, popServDF)
}
\arguments{
\item{waterQuantDF}{data frame}

\item{popServDF}{data frame}
}
\description{
Function to merge Water Quantity and Population Served data frames into one
by "to_agency code", "to_site_no", "Year"
}
\examples{
pathToSample <- system.file("extdata",package = "wateRuseSWUDS")
dp <- readPopServedXL(file.path(pathToSample,"OHpopserved_output.xlsx"))
dq <- readWaterQuantityXL(file.path(pathToSample,
              "OH_CTF_SW_monthly_permit_sample_data.xlsx"))
df <- mergeWaterQuantPopServ(waterQuantDF = dq, popServDF = dp)
}
