% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swuds_parse_functions.R
\name{read_swuds_pop}
\alias{read_swuds_pop}
\title{Function to read in SWUDS Population Served Excel file into a dataframe
and converts column names to NWIS codes using a lookup table nwis_lookup.xlsx}
\usage{
read_swuds_pop(file_path)
}
\arguments{
\item{file_path}{path to file}
}
\description{
Function to read in SWUDS Population Served Excel file into a dataframe
and converts column names to NWIS codes using a lookup table nwis_lookup.xlsx
}
\examples{
pathToSample <- system.file("extdata",package = "wateRuseSWUDS")
dp <- read_swuds_pop(file.path(pathToSample,"OHpopserved_output.xlsx"))
}
