% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swuds_parse_functions.R
\name{readPopServedXL}
\alias{readPopServedXL}
\title{Function to read in SWUDS Population Served Excel file into a dataframe
and converts column names to NWIS codes using a lookup table nwis_lookup.xlsx}
\usage{
readPopServedXL(file_path)
}
\arguments{
\item{file_path}{path to file}
}
\description{
Function to read in SWUDS Population Served Excel file into a dataframe
and converts column names to NWIS codes using a lookup table nwis_lookup.xlsx
}
\examples{
pathToSample <- system.file("extdata",package = "wateRuseSWUDS")
dp <- readPopServedXL(file.path(pathToSample,"OHpopserved_output.xlsx"))
}
