% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swuds_parse_functions.R
\name{readWaterQuantityXL}
\alias{readWaterQuantityXL}
\title{Read Excel file Water Quantity}
\usage{
readWaterQuantityXL(file_path)
}
\arguments{
\item{file_path}{path to Excel file}
}
\value{
data frame
}
\description{
Function to read in SWUDS Water Quantity Excel file into a dataframe
and converts column names to NWIS codes using a lookup table nwis_lookup.xlsx
}
\examples{
pathToSample <- system.file("extdata",package = "wateRuseSWUDS")
df <- readWaterQuantityXL(file.path(pathToSample, "OH_CTF_SW_monthly_permit_sample_data.xlsx"))
}
