% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_swuds.R
\name{create_swuds}
\alias{create_swuds}
\alias{summary.swuds}
\alias{as_swuds}
\title{Load and check swuds data}
\usage{
create_swuds(path_pop, path_quant, ...)

\method{summary}{swuds}(object, ...)

as_swuds(dq, dp = NULL)
}
\arguments{
\item{path_pop}{file path to Excel file of population file}

\item{path_quant}{file path to Excel file of water quantity file}

\item{\dots}{additional paremeters sent to \code{as_swuds}.}

\item{object}{swuds object}

\item{dq}{data frame quanity}

\item{dp}{data frame population}
}
\description{
Load and check swuds data
}
\examples{
path_to_data <-  system.file("extdata", package="wateRuseSWUDS")
file_name_1 <- "OH_CTF_SW_monthly_permit_sample_data.xlsx"
file_name_2 <- "OHpopserved_output.xlsx"
excel_path_quant <- file.path(path_to_data, file_name_1)
excel_path_pop <- file.path(path_to_data, file_name_2)
swuds_1 <- create_swuds(excel_path_pop, excel_path_quant)
summary(swuds_1)
}
