% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_site_monthly.R
\name{time_series_site_monthly}
\alias{time_series_site_monthly}
\title{Monthly Time Series}
\usage{
time_series_site_monthly(s_wuds, FROM_SITE_NUM)
}
\arguments{
\item{s_wuds}{dataframe, the swuds water use data}

\item{FROM_SITE_NUM}{chr, site number agency and site_no unigue to site}
}
\description{
Plot water use monthly (volume_mgd) overlaying multiple years
}
\examples{
s_wuds <- swuds_sample #example data from Ohio
FROM_SITE_NUM <- c("410233083375001")
testp2 <- time_series_site_monthly(s_wuds, FROM_SITE_NUM)
}
