% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wateRuseSWUDS.R
\docType{package}
\name{WUReview-package}
\alias{WUReview-package}
\title{WUReview}
\description{
\tabular{ll}{
Package: \tab WUReview\cr
Type: \tab Package\cr
License: \tab Unlimited for this package, dependencies have more
restrictive licensing.\cr
LazyLoad: \tab yes\cr
}
}
\details{
Collection of functions to help parse and analyze AWUDS data.
}
