% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_aquifier}
\alias{aggregate_aquifier}
\alias{aggregate_county}
\alias{aggregate_HUC}
\alias{aggregate_month}
\alias{aggregate_season}
\alias{aggregate_year}
\title{aggregate_HUC}
\usage{
aggregate_aquifier(s_wuds, aquifer, sum_avg)

aggregate_county(s_wuds, state_county, sum_avg)

aggregate_HUC(s_wuds, HUC, sum_avg)

aggregate_month(s_wuds, month, sum_avg)

aggregate_season(s_wuds, season, sum_avg)

aggregate_year(s_wuds, year, sum_avg)
}
\arguments{
\item{s_wuds}{dataframe, the swuds water use data}

\item{aquifer}{chr, the aquifer of interest to be aggregated}

\item{sum_avg}{chr, user selects either "sum" or "avg" and the sum
or average will be computed}

\item{state_county}{chr, the county of interest to be aggregated.
State abbreviation is also
appended to the beginning of the county}

\item{HUC}{chr, the HUC of interest to be aggregated}

\item{month}{chr, the month of interest to be aggregated in abb
format (i.e. Jan, Feb, etc.)}

\item{season}{chr, the aquifer of interest to be aggregated}

\item{year}{numeric, the year of interest to be aggregated}
}
\description{
Allows user to aggregate their data from SWUDS by user-defined aquifer(s)
}
\examples{
s_wuds <- swuds_sample #example data from Ohio
aquifer <- "TBD"
sum_avg <- "sum"
test_aquifer <- aggregate_aquifier(s_wuds, aquifer, sum_avg)
state_county <- "39033"
test_county <- aggregate_county(s_wuds, state_county, sum_avg)
HUC <- c("041000040104", "050600020105")
test_HUC <- aggregate_HUC(s_wuds, HUC, sum_avg)
month <- c("Jan", "Feb")
test_month <- aggregate_month(s_wuds, month, sum_avg)
season <- "Fall"
test_aggSeason <- aggregate_season(s_wuds, season, sum_avg)
year <- c("1995", "2004")
test_year <- aggregate_year(s_wuds, year, sum_avg)
}
