% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_county.R
\name{aggregate_county}
\alias{aggregate_county}
\title{aggregate_county}
\usage{
aggregate_county(s.wuds, state_county, sum_avg)
}
\arguments{
\item{s.wuds}{dataframe, the swuds water use data}

\item{state_county}{chr, the county of interest to be aggregated. State abbreviation is also
appended to the beginning of the county}

\item{sum_avg}{chr, user selects either "sum" or "avg" and the sum or average will be computed}
}
\description{
Allows user to aggregate their data from SWUDS by user-defined county(ies)
}
\examples{
s.wuds <- swudsSample #example data from Ohio
HUC <- c("041000040104", "050600020105")
sum_avg <- "sum"
test_county <- aggregate_HUC(s.wuds, state_county, sum_avg)

}
