% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_HUC.R
\name{aggregate_HUC}
\alias{aggregate_HUC}
\title{aggregate_HUC}
\usage{
aggregate_HUC(s_wuds, HUC, sum_avg)
}
\arguments{
\item{s_wuds}{dataframe, the swuds water use data}

\item{HUC}{chr, the HUC of interest to be aggregated}

\item{sum_avg}{chr, user selects either "sum" or "avg" and
the sum or average will be computed}
}
\description{
Allows user to aggregate their data from SWUDS by user-defined HUC(S)
}
\examples{
s_wuds <- swuds_sample #example data from Ohio
HUC <- c("041000040104", "050600020105")
sum_avg <- "sum"
test_HUC <- aggregate_HUC(s_wuds, HUC, sum_avg)
}
