% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_swuds.R
\name{explore_swuds}
\alias{explore_swuds}
\title{Explore data in the shiny application}
\usage{
explore_swuds(browse = TRUE)
}
\arguments{
\item{browse}{Logical. Use browser for running Shiny app.}
}
\description{
Open an interactive app in a browser.
}
\examples{
\dontrun{
explore_swuds()
}
}
