% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swuds_parse_functions.R
\name{readPopServedXL}
\alias{readPopServedXL}
\title{Function to read in SWUDS Population Served Excel file into a dataframe
and converts column names to NWIS codes using a lookup table nwis_lookup.xlsx}
\usage{
readPopServedXL(file_path)
}
\description{
Function to read in SWUDS Population Served Excel file into a dataframe
and converts column names to NWIS codes using a lookup table nwis_lookup.xlsx
}
