% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesswuds.R
\name{time_series_data}
\alias{time_series_data}
\title{timeseriesswuds}
\usage{
time_series_data(w.use, data.elements, area.column, plot.points = TRUE,
  years = NA, areas = NA, y.scale = NA, log = FALSE,
  legend = TRUE, c.palette = c("#999999", "#E69F00", "#56B4E9",
  "#009E73", "#F0E442", "#0072B2", "#D55E00", "#CC79A7"))
}
\arguments{
\item{plot.points}{is a logical function to show counties as points or clustered bar graph}

\item{y.scale}{allows R to set the y-axis scale given available data range. Defaults to NA which lets R set the scale based on dataset values.}

\item{log}{= TRUE or FALSE allows user to set log scale, default is FALSE}

\item{legend}{is a logical function to include list of counties in a legend if manageable, default is TRUE}

\item{c.palette}{color palette to use for points or bars}

\item{dec_date}{is decimate date in fractions of year}

\item{volume_mgd}{is water-use rate in million gallons per day, monthly and annual}

\item{categories}{are character categorical attributes defined in your datafile such as 
water-use caragory, state, county, HUC, aquifer, site type, water type, year, month, season, quantile by which data are filtered}

\item{categories.column}{character that defines which column to use to specify category

# old from wateRuse}
}
\description{
timeseriesswuds
}
